Chapter statistical analysis
Created a folder with the scraping results and notebooks
for the analysis. Note that some of the following have
been in the Axelbib repository before. They have been
moved here now because i believe it's more appropiate.

I am working on dividing the analysis on individual
notebooks. The structure is the following:

- General analysis
- Temporal analysis
- Co-authors network analysis
- Clustering
- Summary

Extra notebooks will be deleted once I have moved
all the information/code of interest.