A first draft.
This contains the introduction in a primary level.
Followed by the works of A. Rpoport and R. Axelrod.

The success of Tit for Tat leads on to critism how the strategies that
outperformed it focusing on Nowak's works. Strategies such as
Generous Tit for Tat, Pavlov.

Also memory one strategies and the representation
using finite state machines.