New structure
- Restructured the paper ones again. Now very simple going throw highlight events that took place in
a chronological order. The current sections (of the timeline) are:
  - Origin and (1961-1972)
  - Axelrod’s Tournaments (1981-1984)
  - Further work and Criticism on Computer Tournaments (1984-1993)
  - Evolutionary Dynamics (1987-1999)
  - Modern Approaches (1995-2011)
  - Zero Determinant (2012 - 2015)
  - Current Area (2015 - 2017)

+ Some of the analysis cause I forgot to commit :smile:

Cleaning notebooks.

The code for retrieving data from Axelbib and pushing
articles to Axelbib have been removed from notebooks
and are now scripts.

Also all the code for cleaning the data were brought together
on a notebook (A.0.1. Cleaning data)

Write the analysis

Cleaning the main body of the article

This includes reading through and correcting
the language. As well as adding connecting sentences.

Clear analysis.

Few things left to do. Add 2-3 final sentences and
generate the network plots with legends.