Analysis chapter (#19)
* Re writting analysis.

After a meeting with Vince the structure of the analysis
was determined #15. Currently writing all the analysis.
The following sections has been complinted:

- Data collection
- Pre Analysis
- Contructing the networks
- Collaborative behaviour

Left to do is:

- Influence, which is mainly written just missing the conclusions
- Conclusion

* running notebooks for analysis

* tex files.

For variables and tables

* Last subsection and conclusions

* clears notebooks.

All notebooks 1-8 run. All tables and constants
are manually written to file

* added notes to what needs to be re written.

* remove keywords that where not used.

A list of keywords have been used to collected data
on the prisoner's dilemma. But actually a few of them
did not return a single article.

These words were removed from the table. The
check was done in the notebook 'Extra Check keywords'

* additions to text

This includes:

- a sentence about the number of keywords for the data collection
- sources citation
- example for the levenshtein distance
- Definition of co authorship network

* first re-write without trends are conclusion

* Finish analysis.

Only conclusion left.

* clean tex files

The files autogenerated for the tables and the variables.

* clean images

Removed unwanted plots and commit new ones.

* run notebooks & added time series analysis

* sty file

* add references

* conclusion

* clean the literature